KYLIN-1186 remove the method for long in BitmapCounter.java

+++ core-metadata/src/test/java/org/apache/kylin/measure/bitmap/BitmapSerializerTest.java
@@ -36,7 +36,7 @@ public class BitmapSerializerTest {
         counter.add(1);
         counter.add(3333);
         counter.add("123".getBytes());
-        counter.add((long)123);
+        counter.add(123);
         assertEquals(3, counter.getCount());
 
         ByteBuffer buffer = ByteBuffer.allocate(10 * 1024 * 1024);
